backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.1
Automat==25.4.16
billiard==4.2.2
celery==5.4.0
certifi==2025.10.5
cffi==2.0.0
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.1.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.4
 djangorestframework==3.15.2
 django-cors-headers==4.4.0
 uvicorn==0.32.1
 websockets==13.1
 telethon==1.37.0
 cryptg>=0.4.0  # Acelera criptografia do Telethon usando OpenSSL nativo
 python-decouple==3.8
 psycopg2-binary==2.9.11
 Pillow==12.0.0
 dramatiq[rabbitmq]==1.18.0
 gunicorn==23.0.0
 django-csp==3.7
django-timezone-field==7.1
fastapi==0.121.0
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jiter==0.11.1
kombu==5.5.4
msgpack==1.1.2
google-genai>=1.51.0
packaging==25.0
pdfminer.six==20231228
qrcode[pil]==8.2
pdfplumber==0.11.4
pika==1.3.2
pillow==12.0.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdfium2==4.30.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
redis==5.3.1
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.49.1
Telethon==1.37.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.32.1
vine==5.1.0
watchdog==6.0.0
watchdog_gevent==0.2.1
wcwidth==0.2.14
websockets==13.1
zope.event==6.0
zope.interface==8.0.1